renamed files, added type to fix error
