+++ pages/components/Content/index.tsx
@@ -1,4 +1,4 @@
-import styles from './style.module.css';
+import styles from './styles.module.css';
 
 interface ContentProps {
     children: JSX.Element;
