+++ pages/components/MortgageCalculator/PeriodSelect/styles.module.css
@@ -1,6 +1,6 @@
 .selectorName {
   font-size: 12px;
   margin-bottom: 5px;
-  color: rgba(47,64,73,0.5);
+  color: var(--color-secondary);
   font-weight: 700;
 }
