+++ pages/components/MortgageCalculator/PurchasePriceSlider/styles.module.css
@@ -1,7 +1,7 @@
 .selectorName {
   font-size: 12px;
   margin-bottom: 5px;
-  color: rgba(47,64,73,0.5);
+  color: var(--color-secondary);
   font-weight: 700;
 }
