+++ pages/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next';
-import styles from './style.module.css';
+import styles from './styles.module.css';
 import { MortgageCalculator } from './components/MortgageCalculator';
