+++ styles/globals.css
@@ -1,9 +1,15 @@
+:root {
+  --color-primary: rgba(47,64,73,1);
+  --color-secondary: rgba(47,64,73,0.5);
+  --color-button: rgb(73, 41, 117);
+}
+
 html,
 body {
   padding: 0;
   margin: 0;
   font-family: Inter;
-  color: rgba(47,64,73,1);
+  color: var(--color-primary);
 }
 
 a {
@@ -26,14 +32,7 @@ h3 {
 }
 
 .centered {
-  width: 90vw;
-  max-width: 90vw;
+  width: 88vw;
+  max-width: 88vw;
   margin: 0 auto;
-}
-
-@media (max-width: 767px) {
-  .centered {
-    width: 88vw;
-    max-width: 88vw;
-  }
 }
